+++ app/src/main/java/woowacourse/payments/ui/newcard/NewCardState.kt
@@ -8,7 +8,7 @@ import woowacourse.payments.domain.CardHolderName
 import woowacourse.payments.domain.CardNumber
 import woowacourse.payments.domain.CardPassword
 import woowacourse.payments.ui.common.model.CardUiModel
-import woowacourse.payments.ui.company.model.CompanyUiModel
+import woowacourse.payments.ui.newcard.model.CompanyUiModel
 import java.time.format.DateTimeFormatter
 
 class NewCardState {
