+++ app/src/main/java/woowacourse/payments/ui/newcard/components/CardCompanySelectRow.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.company.components
+package woowacourse.payments.ui.newcard.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -22,8 +22,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.domain.CardCompany
-import woowacourse.payments.ui.company.model.CompanyUiModel
-import woowacourse.payments.ui.company.model.toUiModel
+import woowacourse.payments.ui.newcard.model.CompanyUiModel
+import woowacourse.payments.ui.newcard.model.toUiModel
 
 private const val COLUMN_COUNT = 4
